feat: OptionTile refactor and dev
