+++ lib/components/OptionTile/OptionTileMulti.tsx
@@ -11,11 +11,11 @@ import { IconTick } from './IconTick';
 import {
 	iconContainer,
 	styledCheckbox,
-	styledWrapper,
-	type StyledWrapperProps,
+	styledOptionTile,
+	type StyledOptionTileProps,
 } from './OptionTile.css';
 
-interface OptionTileProps extends AriaCheckboxProps {
+interface OptionTileMultiProps extends AriaCheckboxProps {
 	/**
 	 * The text label for the checkbox
 	 */
@@ -37,32 +37,37 @@ interface OptionTileProps extends AriaCheckboxProps {
 
 const uniformDataAttr = (v: boolean) => (v ? '' : undefined);
 
-const Wrapper = (props: PropsWithChildren<StyledWrapperProps>) => {
+const Wrapper = (props: PropsWithChildren<StyledOptionTileProps>) => {
 	const { disabled, selected, ...restProps } = props;
 	return (
 		<label
-			className={styledWrapper({ disabled, selected })}
+			className={styledOptionTile({ disabled, selected })}
 			{...restProps}
 		/>
 	);
 };
 
 /**
- * The OptionTile is primarily used in the booking flow to provide either single or multi-select optiosn for the user.
+ * The OptionTileMulti is primarily used in the booking flow to provide multi-select options for the user.
+ * It is a wrapper around the native checkbox input, and supports controlled and uncontrolled usage.
+ * For groups of closely related checkboxes, it is often better to have a custom implementation that uses the React
+ * Aria hooks `useCheckboxGroup` and `useCheckbox` that manage accessibility and keyboard intereraction while Overdrive
+ * style recipes provide the appearance.
  */
-export const OptionTile = ({
+export const OptionTileMulti = ({
 	description,
 	icon,
 	indicator = 'checkbox',
 	label,
 	...props
-}: OptionTileProps) => {
+}: OptionTileMultiProps) => {
 	const ref = useRef<HTMLInputElement>(null);
 	const state = useToggleState(props);
 	const { inputProps, isSelected } = useCheckbox(props, state, ref);
 	const { focusProps, isFocusVisible } = useFocusRing();
 	const { hoverProps, isHovered } = useHover({});
-	const showCheckbox = !icon || isHovered || isSelected;
+	const showCheckbox =
+		(!icon || isHovered || isSelected) && indicator === 'checkbox';
 
 	return (
 		<Wrapper
@@ -85,7 +90,7 @@ export const OptionTile = ({
 						<IconTick />
 					</div>
 				) : (
-					<Icon icon={icon} size="100%" />
+					<Icon icon={icon!} size="100%" />
 				)}
 			</div>
 			<div>
